Reduce num iterations for serial python pi
This takes a very long time compared to the C version, and so
we reduce the number of iterations to 1 million. The value of pi
is still accurate to the same decimal place that is printed by the
C version. We also print out a warning of this effect for user info.
Fixes #19.