Update instructions for how to alter device type
All the C examples now run on the CL_DEVICE_TYPE_DEFAULT.
This can be changed in the source files as before, or in the
individual example Makefiles. Also, for an easy change, it can be
changed in the global Makefiles - and this is propogated as expected
to the examples themselves when they are built.

This fixes #20.