Include Acorn example
Acorn is a Methuselah, and is stable after 5206 generations.
We only look at a small number of this as we need (with the
torodial array) to keep track of the 13 escaped gliders which
could get a long way away. If we find the right size box we can try
this.