Very rough implementation of a viewer for life
Compile with -lGL and -lglut.
Usage: ./a.out input.dat input.params
The dat is displayed as red squares (5x5 pixels) on a black background.
Not sure if all the code is necessary - it is certainly very messy!
Aim: nice to actually see the life patterns in non-ascii. Usefull for
very large patterns (i.e. Acorn)